Add example request and book validators

+++ src/controller/bookController.js
@@ -1,28 +1,55 @@
 import { Book } from '../models/Book.js'
 import mongoose from 'mongoose'
+import { validationResult } from 'express-validator'
 
 export const addBook = (req, res) => {
-  const name = req.body.name
-  const author = req.body.author
-  const description = req.body.description
+  const error = validationResult(req)
 
-  const book = new Book({
-    name,
-    author,
-    description
-  })
+  if (!error.isEmpty()) {
+    return res.status(400).send({
+      success: false,
+      errors: error.array()
+    })
+  }
+
+  const { name, author, description } = req.body
+  const book = new Book({ name, author, description })
 
   book.save()
-  return book
+  return res.send({
+    success: true,
+    book
+  })
 }
 
 export const removeBook = async (req, res) => {
+  const error = validationResult(req)
+
+  if (!error.isEmpty()) {
+    return res.status(400).send({
+      success: false,
+      errors: error.array()
+    })
+  }
+
   const bookId = req.query.bookId
   if (mongoose.Types.ObjectId.isValid(bookId)) {
     const book = await Book.findByIdAndDelete(bookId)
-    return book ? res.send(`Book with id ${bookId} deleted`) : res.status(404).send('Book not found')
+    return book
+      ? res.send({
+        success: true,
+        msg: `Book with id ${bookId} deleted`
+      })
+      : res.status(404).send({
+        success: false,
+        errors: 'Book not found'
+      })
   } else {
-    return res.status(400).send('Invalid book id')
+    return res.status(400).send({
+      success: false,
+      errors: 'Invalid book id'
+    }
+    )
   }
 }
 
@@ -31,16 +58,32 @@ export const updateBook = async (req, res) => {
   const description = req.body.description
 
   if (!author && !description) {
-    return res.status(400).send('Specify author or description to update')
+    return res.status(400).send({
+      success: false,
+      errors: 'Specify author or description to update'
+    })
   }
 
   const bookId = req.query.bookId
   if (mongoose.Types.ObjectId.isValid(bookId)) {
     const book = await Book.findByIdAndUpdate(bookId, { author, description }, { returnDocument: 'after' })
 
-    return book ? res.send(book) : res.status(404).send('Book not found')
+    return book
+      ? res.send({
+        success: true,
+        book
+      })
+      : res.status(404).send(
+        {
+          success: false,
+          errors: 'Book not found'
+        })
   } else {
-    return res.send('Invalid book id')
+    return res.status(400).send({
+      success: false,
+      errors: 'Invalid book id'
+    }
+    )
   }
 }
